Move from ldns to a custom stack-based DNS implementation
Due to many heap allocations, ldns has imposed a major performance penalty. Thus, a minimalistic custom DNS implementation is supposed to replace ldns.
The version on this branch is supposed to eventually replace the master branch. It makes use of epoll and is no longer busy polling by default.

First tests have shown that a 4x speedup can be achieved compared to the old massdns version. The goal of this branch is the implementation of fully network-bound stub resolver.