ndjson output add ttl value
The TTL value has a certain reference role in judging whether the subdomain is pan-parsed in the scene of subdomain enumeration. It is recommended that the ttl value be written to the json output result by default. For details, please refer to:http://sh3ll.me/archives/201704041222.txt

Thank you for open source this great tool, I very much hope that the output will have a TTL value.
Note: Compiled and tested without problems.